@@ -24,30 +24,30 @@
     "@hembio/babel-config": "workspace:packages/configs/babel",
     "@hembio/eslint-config": "workspace:packages/configs/eslint",
     "@hembio/tsconfig": "workspace:packages/configs/typescript",
-    "@types/eslint": "^7.28.0",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^16.4.13",
-    "@types/react": "^17.0.16",
-    "@types/rimraf": "^3.0.1",
-    "@types/tar-stream": "^2.2.1",
-    "@yarnpkg/pnpify": "^3.0.0-rc.14",
+    "@types/eslint": "^8.4.2",
+    "@types/jest": "^27.5.1",
+    "@types/node": "^17.0.35",
+    "@types/react": "^18.0.9",
+    "@types/rimraf": "^3.0.2",
+    "@types/tar-stream": "^2.2.2",
+    "@yarnpkg/pnpify": "^4.0.0-rc.6",
     "cross-env": "^7.0.3",
-    "depcheck": "^1.4.2",
-    "eslint": "^7.32.0",
+    "depcheck": "^1.4.3",
+    "eslint": "^8.16.0",
     "execa": "^5.1.1",
-    "globby": "^12.0.0",
-    "husky": "^7.0.1",
-    "jest": "^27.0.6",
-    "jest-watch-typeahead": "^0.6.4",
-    "lint-staged": "^11.1.2",
-    "node-gyp": "^8.1.0",
-    "prettier": "^2.3.2",
+    "globby": "^12.2.0",
+    "husky": "^8.0.1",
+    "jest": "^28.1.0",
+    "jest-watch-typeahead": "^1.1.0",
+    "lint-staged": "^12.4.1",
+    "node-gyp": "^9.0.0",
+    "prettier": "^2.6.2",
     "regenerator-runtime": "^0.13.9",
     "rimraf": "^3.0.2",
     "tar-stream": "^2.2.0",
-    "ts-node": "^10.1.0",
-    "tslib": "^2.3.0",
-    "typescript": "^4.3.5"
+    "ts-node": "^10.8.0",
+    "tslib": "^2.4.0",
+    "typescript": "^4.6.4"
   },
   "resolutions": {
     "@types/eslint": "latest",
@@ -64,7 +64,7 @@
     "react": "latest",
     "react-dom": "latest",
     "react-refresh": "latest",
-    "graphql": "latest",
+    "graphql": "^15",
     "fastify": "^3.18.1",
     "fastify-cors": "^6.0.1",
     "fastify-formbody": "^5.0.0",
@@ -92,5 +92,5 @@
     ],
     "*.{md,html,css,yml,yaml}": "prettier --write"
   },
-  "packageManager": "yarn@3.0.0"
+  "packageManager": "yarn@3.2.1"
 }
