@@ -21,42 +21,41 @@
     "forge": "forge.config.js"
   },
   "devDependencies": {
-    "@electron-forge/cli": "^6.0.0-beta.59",
-    "@electron-forge/maker-deb": "^6.0.0-beta.59",
-    "@electron-forge/maker-rpm": "^6.0.0-beta.59",
-    "@electron-forge/maker-squirrel": "^6.0.0-beta.59",
-    "@electron-forge/maker-wix": "^6.0.0-beta.59",
-    "@electron-forge/maker-zip": "^6.0.0-beta.59",
-    "@electron-forge/plugin-webpack": "6.0.0-beta.59",
-    "@electron-forge/shared-types": "^6.0.0-beta.59",
+    "@electron-forge/cli": "^6.0.0-beta.63",
+    "@electron-forge/maker-deb": "^6.0.0-beta.63",
+    "@electron-forge/maker-rpm": "^6.0.0-beta.63",
+    "@electron-forge/maker-squirrel": "^6.0.0-beta.63",
+    "@electron-forge/maker-wix": "^6.0.0-beta.63",
+    "@electron-forge/maker-zip": "^6.0.0-beta.63",
+    "@electron-forge/plugin-webpack": "6.0.0-beta.63",
+    "@electron-forge/shared-types": "^6.0.0-beta.63",
     "@marshallofsound/webpack-asset-relocator-loader": "^0.5.0",
-    "@types/eslint": "^7.28.0",
-    "@types/prop-types": "^15.7.4",
-    "@types/react": "^17.0.16",
-    "@types/webpack-node-externals": "^2.5.2",
-    "@typescript-eslint/eslint-plugin": "^4.29.0",
-    "@typescript-eslint/parser": "^4.29.0",
-    "copy-webpack-plugin": "^9.0.1",
+    "@types/eslint": "^8.4.2",
+    "@types/prop-types": "^15.7.5",
+    "@types/react": "^18.0.9",
+    "@types/webpack-node-externals": "^2.5.3",
+    "@typescript-eslint/eslint-plugin": "^5.25.0",
+    "@typescript-eslint/parser": "^5.25.0",
+    "copy-webpack-plugin": "^9.1.0",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.2.0",
-    "electron": "^13.1.8",
+    "css-loader": "^6.7.1",
+    "electron": "^13.6.9",
     "electron-wix-msi": "^3.2.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-import": "^2.23.4",
+    "eslint": "^8.16.0",
+    "eslint-plugin-import": "^2.26.0",
     "file-loader": "^6.2.0",
-    "fork-ts-checker-webpack-plugin": "^6.3.1",
+    "fork-ts-checker-webpack-plugin": "^6.5.2",
     "node-loader": "^2.0.0",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "style-loader": "^3.2.1",
-    "ts-loader": "^9.2.5",
-    "typescript": "^4.3.5",
-    "webpack": "^5.49.0",
+    "prop-types": "^15.8.1",
+    "react": "^18.1.0",
+    "style-loader": "^3.3.1",
+    "ts-loader": "^9.3.0",
+    "typescript": "^4.6.4",
+    "webpack": "^5.72.1",
     "webpack-node-externals": "^3.0.0"
   },
   "dependencies": {
-    "electron-squirrel-startup": "^1.0.0",
-    "exe-icon-extractor": "github:bitdisaster/exe-icon-extractor"
+    "electron-squirrel-startup": "^1.0.0"
   },
   "installConfig": {
     "hoistingLimits": "dependencies"
