@@ -12,19 +12,19 @@
     "@hembio/core": "workspace:packages/libs/core",
     "@hembio/logger": "workspace:packages/libs/logger",
     "chalk": "^4.1.2",
-    "fastify": "^3.20.1",
-    "fastify-static": "^4.2.3",
-    "finalhandler": "^1.1.2",
+    "fastify": "^3.29.0",
+    "fastify-static": "^4.7.0",
+    "finalhandler": "^1.2.0",
     "hostile": "^1.3.3",
     "http2-proxy": "^5.0.53",
-    "http2-wrapper": "^2.1.2",
+    "http2-wrapper": "^2.1.11",
     "internal-ip": "^6.2.0"
   },
   "devDependencies": {
     "@types/finalhandler": "^1.1.1",
     "@types/hostile": "^1.3.0",
     "cross-env": "^7.0.3",
-    "ts-node": "^10.1.0",
-    "typescript": "^4.3.5"
+    "ts-node": "^10.8.0",
+    "typescript": "^4.6.4"
   }
 }
