@@ -19,81 +19,81 @@
   "license": "SEE LICENSE IN LICENSE",
   "dependencies": {
     "$": "link:public",
-    "@apollo/client": "^3.4.5",
-    "@emotion/react": "^11.4.0",
-    "@emotion/styled": "^11.3.0",
+    "@apollo/client": "^3.6.4",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
     "@hembio/core": "workspace:packages/libs/core",
     "@hembio/logger": "workspace:packages/libs/logger",
-    "@material-ui/core": "^5.0.0-beta.2",
-    "@material-ui/icons": "^5.0.0-beta.1",
-    "@material-ui/lab": "^5.0.0-alpha.41",
-    "@material-ui/styles": "^5.0.0-beta.2",
-    "@material-ui/system": "5.0.0-beta.2",
-    "axios": "^0.21.1",
-    "class-validator": "^0.13.1",
+    "@mui/icons-material": "^5.8.0",
+    "@mui/lab": "^5.0.0-alpha.82",
+    "@mui/material": "^5.8.0",
+    "@mui/styles": "^5.8.0",
+    "@mui/system": "^5.8.0",
+    "axios": "^0.27.2",
+    "class-validator": "^0.13.2",
     "clsx": "^1.1.1",
     "cookie-storage": "^6.1.0",
-    "graphql": "^15.5.1",
-    "hls.js": "^1.0.7",
-    "i18next": "^20.3.5",
-    "i18next-browser-languagedetector": "^6.1.2",
+    "graphql": "^15.8.0",
+    "hls.js": "^1.1.5",
+    "i18next": "^20.6.1",
+    "i18next-browser-languagedetector": "^6.1.4",
     "i18next-chained-backend": "^3.0.2",
     "i18next-localstorage-cache": "^1.1.1",
     "i18next-resources-to-backend": "^1.0.0",
-    "mobx": "^6.3.2",
-    "mobx-react-lite": "^3.2.0",
-    "react": "^17.0.2",
-    "react-apollo-network-status": "^5.0.1",
-    "react-dom": "^17.0.2",
-    "react-i18next": "^11.11.4",
+    "mobx": "^6.5.0",
+    "mobx-react-lite": "^3.4.0",
+    "react": "^18.1.0",
+    "react-apollo-network-status": "^5.1.0",
+    "react-dom": "^18.1.0",
+    "react-i18next": "^11.16.9",
     "react-infinite-scroll-component": "^6.1.0",
-    "react-router": "^5.2.0",
+    "react-router": "^6.3.0",
     "react-router-config": "^5.1.1",
-    "react-router-dom": "^5.2.0",
-    "react-syntax-highlighter": "^15.4.4",
+    "react-router-dom": "^6.3.0",
+    "react-syntax-highlighter": "^15.5.0",
     "react-transition-group": "^4.4.2",
-    "react-useanimations": "^2.0.6",
-    "react-virtual": "^2.8.1",
-    "react-window": "^1.8.6",
+    "react-useanimations": "^2.0.8",
+    "react-virtual": "^2.10.4",
+    "react-window": "^1.8.7",
     "~": "link:src"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.0",
-    "@babel/plugin-proposal-class-properties": "^7.14.5",
-    "@babel/plugin-proposal-decorators": "^7.14.5",
-    "@babel/preset-env": "^7.15.0",
-    "@babel/preset-typescript": "^7.15.0",
+    "@babel/core": "^7.18.0",
+    "@babel/plugin-proposal-class-properties": "^7.17.12",
+    "@babel/plugin-proposal-decorators": "^7.17.12",
+    "@babel/preset-env": "^7.18.0",
+    "@babel/preset-typescript": "^7.17.12",
     "@createnextapp/async-local-storage": "^1.0.1",
     "@hembio/eslint-config": "workspace:packages/configs/eslint",
     "@hembio/tsconfig": "workspace:packages/configs/typescript",
-    "@testing-library/dom": "^8.1.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.2.1",
-    "@types/babel__core": "^7.1.15",
+    "@testing-library/dom": "^8.13.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^14.2.0",
+    "@types/babel__core": "^7.1.19",
     "@types/babel__preset-env": "^7.9.2",
-    "@types/eslint": "^7.28.0",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^16.4.13",
-    "@types/prettier": "^2.3.2",
-    "@types/react": "^17.0.16",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router": "^5.1.16",
-    "@types/react-router-config": "^5.0.3",
-    "@types/react-router-dom": "^5.1.8",
-    "@types/react-syntax-highlighter": "^13.5.2",
-    "@types/react-transition-group": "^4.4.2",
+    "@types/eslint": "^8.4.2",
+    "@types/jest": "^27.5.1",
+    "@types/node": "^17.0.35",
+    "@types/prettier": "^2.6.1",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
+    "@types/react-router": "^5.1.18",
+    "@types/react-router-config": "^5.0.6",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-syntax-highlighter": "^15.5.1",
+    "@types/react-transition-group": "^4.4.4",
     "@types/react-window": "^1.8.5",
     "@types/webpack": "^5.28.0",
-    "@typescript-eslint/eslint-plugin": "^4.29.0",
-    "@typescript-eslint/parser": "^4.29.0",
-    "@vitejs/plugin-react-refresh": "^1.3.6",
+    "@typescript-eslint/eslint-plugin": "^5.25.0",
+    "@typescript-eslint/parser": "^5.25.0",
+    "@vitejs/plugin-react": "^1.3.2",
     "babel-eslint": "^10.1.0",
     "babel-plugin-transform-metadata": "^4.0.2",
-    "jest": "^27.0.6",
-    "typescript": "^4.3.5",
-    "vite": "^2.4.4",
-    "yarn": "^1.22.11"
+    "jest": "^28.1.0",
+    "typescript": "^4.6.4",
+    "vite": "^2.9.9",
+    "yarn": "^1.22.18"
   },
   "scripts": {
     "dev": "vite",
