@@ -4,14 +4,14 @@
   "description": "Sharable Webpack Configurations",
   "dependencies": {
     "run-script-webpack-plugin": "^0.0.11",
-    "ts-loader": "^9.2.5",
-    "typescript": "^4.3.5",
-    "webpack": "^5.49.0",
-    "webpack-cli": "^4.7.2",
+    "ts-loader": "^9.3.0",
+    "typescript": "^4.6.4",
+    "webpack": "^5.72.1",
+    "webpack-cli": "^4.9.2",
     "webpack-node-externals": "^3.0.0"
   },
   "devDependencies": {
-    "@types/webpack-node-externals": "^2.5.2",
+    "@types/webpack-node-externals": "^2.5.3",
     "webpack-merge": "^5.8.0"
   }
 }
