@@ -20,14 +20,14 @@
     "execa": "^5.1.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.0",
-    "@babel/preset-env": "^7.15.0",
-    "@babel/preset-typescript": "^7.15.0",
-    "@types/babel__core": "^7.1.15",
+    "@babel/core": "^7.18.0",
+    "@babel/preset-env": "^7.18.0",
+    "@babel/preset-typescript": "^7.17.12",
+    "@types/babel__core": "^7.1.19",
     "@types/babel__preset-env": "^7.9.2",
-    "@types/jest": "^26.0.24",
-    "jest": "^27.0.6",
-    "ts-node": "^10.1.0",
-    "typescript": "^4.3.5"
+    "@types/jest": "^27.5.1",
+    "jest": "^28.1.0",
+    "ts-node": "^10.8.0",
+    "typescript": "^4.6.4"
   }
 }
