@@ -30,54 +30,54 @@
     "codegen": "ts-node --transpile-only ./src/codegen.ts"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.0",
-    "@babel/preset-env": "^7.15.0",
-    "@babel/preset-typescript": "^7.15.0",
-    "@graphql-codegen/add": "^3.0.0",
-    "@graphql-codegen/cli": "^2.0.1",
-    "@graphql-codegen/typescript": "^2.0.0",
-    "@graphql-codegen/typescript-operations": "^2.0.1",
-    "@graphql-codegen/typescript-react-apollo": "^3.0.0",
+    "@babel/core": "^7.18.0",
+    "@babel/preset-env": "^7.18.0",
+    "@babel/preset-typescript": "^7.17.12",
+    "@graphql-codegen/add": "^3.1.1",
+    "@graphql-codegen/cli": "^2.6.2",
+    "@graphql-codegen/typescript": "^2.4.11",
+    "@graphql-codegen/typescript-operations": "^2.4.0",
+    "@graphql-codegen/typescript-react-apollo": "^3.2.14",
     "@hembio/tsconfig": "workspace:packages/configs/typescript",
     "@hembio/webpack-config": "workspace:packages/configs/webpack",
     "@nestjs/cli": "^7.6.0",
-    "@types/babel__core": "^7.1.15",
+    "@types/babel__core": "^7.1.19",
     "@types/babel__preset-env": "^7.9.2",
     "@types/bcrypt": "^5.0.0",
-    "@types/body-parser": "^1.19.1",
-    "@types/cookie-parser": "^1.4.2",
+    "@types/body-parser": "^1.19.2",
+    "@types/cookie-parser": "^1.4.3",
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
     "@types/express-session": "^1.17.4",
-    "@types/jest": "^26.0.24",
-    "@types/jsonwebtoken": "^8.5.4",
-    "@types/qrcode": "^1.4.1",
+    "@types/jest": "^27.5.1",
+    "@types/jsonwebtoken": "^8.5.8",
+    "@types/qrcode": "^1.4.2",
     "@types/range-parser": "^1.2.4",
-    "@types/rimraf": "^3.0.1",
-    "@types/sql.js": "^1.4.2",
+    "@types/rimraf": "^3.0.2",
+    "@types/sql.js": "^1.4.3",
     "@types/subtitle": "^2.0.3",
-    "cookie-parser": "^1.4.5",
+    "cookie-parser": "^1.4.6",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.2.0",
+    "css-loader": "^6.7.1",
     "file-loader": "^6.2.0",
     "html-loader": "^2.1.2",
-    "jest": "^27.0.6",
+    "jest": "^28.1.0",
     "json-loader": "^0.5.7",
-    "nexe": "^4.0.0-beta.18",
+    "nexe": "^4.0.0-rc.1",
     "nexe-natives": "^1.0.2",
-    "pkg": "^5.3.1",
+    "pkg": "^5.7.0",
     "require-glob": "^4.0.0",
     "rimraf": "^3.0.2",
-    "ts-loader": "^9.2.5",
-    "ts-node": "^10.1.0",
-    "typescript": "^4.3.5",
-    "webpack": "^5.49.0",
-    "webpack-cli": "^4.7.2",
+    "ts-loader": "^9.3.0",
+    "ts-node": "^10.8.0",
+    "typescript": "^4.6.4",
+    "webpack": "^5.72.1",
+    "webpack-cli": "^4.9.2",
     "~": "link:src"
   },
   "dependencies": {
     "@ffmpeg-installer/ffmpeg": "^1.1.0",
-    "@ffprobe-installer/ffprobe": "^1.2.0",
+    "@ffprobe-installer/ffprobe": "^1.4.1",
     "@hapi/accept": "^5.0.2",
     "@hembio/core": "workspace:packages/libs/core",
     "@hembio/discovery": "workspace:packages/libs/discovery",
