@@ -1,4 +1,4 @@
-import { createStyles, makeStyles } from "@material-ui/styles";
+import { createStyles, makeStyles } from "@mui/styles";
 import { useEffect, useRef } from "react";
 import { Aurora } from "./Aurora";
 import { Curve } from "./Curve";
