@@ -1,14 +1,14 @@
-import Fab from "@material-ui/core/Fab";
-import Grid from "@material-ui/core/Grid";
-import Menu from "@material-ui/core/Menu";
-import MenuItem from "@material-ui/core/MenuItem";
-import Zoom from "@material-ui/core/Zoom";
-import { red } from "@material-ui/core/colors";
-import FavoriteIcon from "@material-ui/icons/Favorite";
-import FavoriteBorderIcon from "@material-ui/icons/FavoriteBorder";
-import MoreHorizIcon from "@material-ui/icons/MoreHoriz";
-import PlayArrowIcon from "@material-ui/icons/PlayArrow";
-import { makeStyles, createStyles } from "@material-ui/styles";
+import Fab from "@mui/material/Fab";
+import Grid from "@mui/material/Grid";
+import Menu from "@mui/material/Menu";
+import MenuItem from "@mui/material/MenuItem";
+import Zoom from "@mui/material/Zoom";
+import { red } from "@mui/material/colors";
+import FavoriteIcon from "@mui/icons-material/Favorite";
+import FavoriteBorderIcon from "@mui/icons-material/FavoriteBorder";
+import MoreHorizIcon from "@mui/icons-material/MoreHoriz";
+import PlayArrowIcon from "@mui/icons-material/PlayArrow";
+import { makeStyles, createStyles } from "@mui/styles";
 import clsx from "clsx";
 import { useState } from "react";
 import { Link } from "react-router-dom";
