@@ -1,16 +1,16 @@
-import AccountBalanceIcon from "@material-ui/icons/AccountBalance";
-import ColorLensIcon from "@material-ui/icons/ColorLens";
-import ContentCutIcon from "@material-ui/icons/ContentCut";
-import CreateIcon from "@material-ui/icons/Create";
-import FaceIcon from "@material-ui/icons/Face";
-import FlareIcon from "@material-ui/icons/Flare";
-import HandymanIcon from "@material-ui/icons/Handyman";
-import MovieCreationIcon from "@material-ui/icons/MovieCreation";
-import MusicNoteIcon from "@material-ui/icons/MusicNote";
-import PersonIcon from "@material-ui/icons/Person";
-import StarIcon from "@material-ui/icons/Star";
-import VideocamIcon from "@material-ui/icons/Videocam";
-import WbIncandescentIcon from "@material-ui/icons/WbIncandescent";
+import AccountBalanceIcon from "@mui/icons-material/AccountBalance";
+import ColorLensIcon from "@mui/icons-material/ColorLens";
+import ContentCutIcon from "@mui/icons-material/ContentCut";
+import CreateIcon from "@mui/icons-material/Create";
+import FaceIcon from "@mui/icons-material/Face";
+import FlareIcon from "@mui/icons-material/Flare";
+import HandymanIcon from "@mui/icons-material/Handyman";
+import MovieCreationIcon from "@mui/icons-material/MovieCreation";
+import MusicNoteIcon from "@mui/icons-material/MusicNote";
+import PersonIcon from "@mui/icons-material/Person";
+import StarIcon from "@mui/icons-material/Star";
+import VideocamIcon from "@mui/icons-material/Videocam";
+import WbIncandescentIcon from "@mui/icons-material/WbIncandescent";
 import { memo } from "react";
 import { CreditListitemProps } from "./CreditByTitleListItem";
 
