@@ -1,12 +1,12 @@
 import { TitleGenreLiterals } from "@hembio/core/src";
-import List from "@material-ui/core/List";
-import ListItem from "@material-ui/core/ListItem";
-import ListItemAvatar from "@material-ui/core/ListItemAvatar";
-import ListItemButton from "@material-ui/core/ListItemButton";
-import ListItemText from "@material-ui/core/ListItemText";
-import Paper from "@material-ui/core/Paper";
-import Skeleton from "@material-ui/core/Skeleton";
-import { experimentalStyled as styled } from "@material-ui/core/styles";
+import List from "@mui/material/List";
+import ListItem from "@mui/material/ListItem";
+import ListItemAvatar from "@mui/material/ListItemAvatar";
+import ListItemButton from "@mui/material/ListItemButton";
+import ListItemText from "@mui/material/ListItemText";
+import Paper from "@mui/material/Paper";
+import Skeleton from "@mui/material/Skeleton";
+import { experimentalStyled as styled } from "@mui/material/styles";
 import { RefObject, useRef } from "react";
 import { useTranslation } from "react-i18next";
 import { Link } from "react-router-dom";
