@@ -1,12 +1,12 @@
 import { TitleGenreSlugs } from "@hembio/core";
-import Box from "@material-ui/core/Box";
-import Checkbox from "@material-ui/core/Checkbox";
-import FormControlLabel from "@material-ui/core/FormControlLabel";
-import IconButton from "@material-ui/core/IconButton";
-import Menu from "@material-ui/core/Menu";
-import Slider from "@material-ui/core/Slider";
-import Typography from "@material-ui/core/Typography";
-import FilterList from "@material-ui/icons/FilterList";
+import Box from "@mui/material/Box";
+import Checkbox from "@mui/material/Checkbox";
+import FormControlLabel from "@mui/material/FormControlLabel";
+import IconButton from "@mui/material/IconButton";
+import Menu from "@mui/material/Menu";
+import Slider from "@mui/material/Slider";
+import Typography from "@mui/material/Typography";
+import FilterList from "@mui/icons-material/FilterList";
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { FilterParams } from "~/containers/TitleList";
