@@ -1,11 +1,11 @@
-import { Tooltip } from "@material-ui/core";
-import ListItemIcon from "@material-ui/core/ListItemIcon";
-import ListItemText from "@material-ui/core/ListItemText";
-import Menu from "@material-ui/core/Menu";
-import MenuItem from "@material-ui/core/MenuItem";
-import MenuList from "@material-ui/core/MenuList";
-import Check from "@material-ui/icons/Check";
-import SubtitlesIcon from "@material-ui/icons/Subtitles";
+import Check from "@mui/icons-material/Check";
+import SubtitlesIcon from "@mui/icons-material/Subtitles";
+import { Tooltip } from "@mui/material";
+import ListItemIcon from "@mui/material/ListItemIcon";
+import ListItemText from "@mui/material/ListItemText";
+import Menu from "@mui/material/Menu";
+import MenuItem from "@mui/material/MenuItem";
+import MenuList from "@mui/material/MenuList";
 import { Observer } from "mobx-react-lite";
 import { memo, useEffect, useState } from "react";
 import { ControlButton } from "./ControlButton";
