@@ -1,25 +1,23 @@
 import { Observer } from "mobx-react-lite";
-import { Route, Redirect, RouteProps } from "react-router-dom";
+import { Navigate } from "react-router-dom";
 import { useStores } from "../stores";
 
-interface ProtectedRouteProps extends RouteProps {
-  default?: boolean;
-  component?: React.ComponentType<RouteProps>;
-}
-
 export const ProtectedRoute = ({
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  component: Component,
   children,
-  ...rest
-}: ProtectedRouteProps): JSX.Element => {
+}: {
+  children: React.ReactNode;
+}): JSX.Element => {
   const { authStore } = useStores();
   return (
     <Observer>
       {() => (
-        <Route {...rest}>
-          {!authStore.isAuthenticated ? <Redirect to="/sign-in" /> : children}
-        </Route>
+        <>
+          {!authStore.isAuthenticated ? (
+            <Navigate to="/sign-in" replace />
+          ) : (
+            children
+          )}
+        </>
       )}
     </Observer>
   );
