@@ -1,5 +1,5 @@
-import Container from "@material-ui/core/Container";
-import { useParams } from "react-router";
+import Container from "@mui/material/Container";
+import { useParams } from "react-router-dom";
 
 export function Credits(): JSX.Element {
   const { titleId } = useParams<{ titleId: string }>();
