@@ -1,16 +1,16 @@
-import { Theme } from "@material-ui/core";
-import Alert from "@material-ui/core/Alert";
-import AlertTitle from "@material-ui/core/AlertTitle";
-import Box from "@material-ui/core/Box";
-import Card from "@material-ui/core/Card";
-import CardMedia from "@material-ui/core/CardMedia";
-import Container from "@material-ui/core/Container";
-import Grid from "@material-ui/core/Grid";
-import Paper from "@material-ui/core/Paper";
-import Skeleton from "@material-ui/core/Skeleton";
-import Typography from "@material-ui/core/Typography";
-import PersonIcon from "@material-ui/icons/Person";
-import { makeStyles, createStyles } from "@material-ui/styles";
+import { Theme } from "@mui/material";
+import Alert from "@mui/material/Alert";
+import AlertTitle from "@mui/material/AlertTitle";
+import Box from "@mui/material/Box";
+import Card from "@mui/material/Card";
+import CardMedia from "@mui/material/CardMedia";
+import Container from "@mui/material/Container";
+import Grid from "@mui/material/Grid";
+import Paper from "@mui/material/Paper";
+import Skeleton from "@mui/material/Skeleton";
+import Typography from "@mui/material/Typography";
+import PersonIcon from "@mui/icons-material/Person";
+import { makeStyles, createStyles } from "@mui/styles";
 import { useParams } from "react-router-dom";
 import { NotFound } from "./NotFound";
 import { CreditByTitleListItem } from "~/components/CreditByTitleListItem";
