@@ -1,13 +1,13 @@
-import { Paper } from "@material-ui/core";
-import Box from "@material-ui/core/Box";
-import Container from "@material-ui/core/Container";
-import Link from "@material-ui/core/Link";
-import Typography from "@material-ui/core/Typography";
-import { Theme } from "@material-ui/core/styles";
-import { makeStyles, createStyles } from "@material-ui/styles";
+import { Paper } from "@mui/material";
+import Box from "@mui/material/Box";
+import Container from "@mui/material/Container";
+import Link from "@mui/material/Link";
+import Typography from "@mui/material/Typography";
+import { Theme } from "@mui/material/styles";
+import { makeStyles, createStyles } from "@mui/styles";
 import clsx from "clsx";
 import { observer } from "mobx-react-lite";
-import { Redirect } from "react-router-dom";
+import { Navigate } from "react-router-dom";
 import { AuroraCanvas } from "~/aurora/AuroraCanvas";
 import { Logo } from "~/components/Logo";
 import { SignInForm } from "~/forms/SignInForm";
@@ -67,7 +67,7 @@ export const SignIn = observer(() => {
   const classes = useStyles();
 
   if (authStore.isAuthenticated) {
-    return <Redirect to="/" />;
+    return <Navigate to="/" replace />;
   }
 
   return (
