@@ -1,5 +1,5 @@
-import { red } from "@material-ui/core/colors";
-import { createTheme } from "@material-ui/core/styles";
+import { red } from "@mui/material/colors";
+import { createTheme } from "@mui/material/styles";
 import { isElectron } from "./utils/isElectron";
 import CocosignumCorsivoItalicoBoldWoff from "$/font.woff";
 import CocosignumCorsivoItalicoBoldWoff2 from "$/font.woff2";
