@@ -1,4 +1,4 @@
-import { useLocation } from "react-router";
+import { useLocation } from "react-router-dom";
 
 export function useQueryString(name: string): string | null {
   const location = useLocation();
