@@ -1,7 +1,7 @@
-import { useParams } from "react-router";
+import { useParams } from "react-router-dom";
 import { TitleList } from "../containers/TitleList";
 
 export const Library = (): JSX.Element => {
-  const { libraryId } = useParams<{ libraryId: string; page: string }>();
+  const { libraryId = "" } = useParams<{ libraryId: string; page: string }>();
   return <TitleList libraryId={libraryId} />;
 };
