@@ -1,5 +1,5 @@
 import path from "path";
-import reactRefresh from "@vitejs/plugin-react-refresh";
+import react from "@vitejs/plugin-react";
 import { defineConfig } from "vite";
 
 export default defineConfig({
@@ -7,17 +7,20 @@ export default defineConfig({
     jsxFactory: `jsx`,
     jsxInject: `import { jsx } from "@emotion/react"; import React from "react";`,
   },
-  plugins: [reactRefresh()],
+  plugins: [react()],
   root: path.resolve(__dirname),
   build: {
     outDir: "./dist",
   },
   server: {
+    host: "hembio.local",
     port: 3000,
     open: false,
     base: "https://hembio.local:3443",
     hmr: {
-      port: 3443,
+      clientPort: 3443,
+      port: 4443,
+      host: "hembio.local",
     },
     fs: {
       allow: [path.resolve("../../../")],
